DABOM-395 feat: 폰트 사이즈 수정
