+++ packages/shared/src/index.ts
@@ -2,13 +2,12 @@ export * from './utils/cn';
 export * from './utils/formatPhoneNumber';
 export * from './utils/formatSize';
 export * from './utils/http';
-export * from './assets/icons';
-export * from './components/Badge';
 export * from './utils/sseClient';
 export { default as Badge } from './components/Badge';
 export { default as Button } from './components/Button';
 export { default as Card } from './components/Card';
 export { default as DaboIcon } from './components/DaboIcon';
+export { default as Drawer } from './components/Drawer';
 export { default as DropDown } from './components/DropDown';
 export { default as Grade } from './components/Grade';
 export { default as Input } from './components/Input';
