+++ packages/shared/src/styles/globals.css
@@ -93,8 +93,13 @@
   line-height: 150%;
   font-weight: 500;
 }
+@utility text-body3-d {
+  font-size: 0.875rem;
+  line-height: 150%;
+  font-weight: 500;
+}
 @utility text-caption-d {
-  font-size: 0.8125rem;
+  font-size: 0.625rem;
   line-height: 140%;
   font-weight: 500;
 }
